refactor : pendingadmininfo 저장하는 함수 클래스 이동(#864)
