+++ src/main/java/com/clubber/ClubberServer/domain/admin/implement/AdminAppender.java
@@ -2,9 +2,6 @@ package com.clubber.ClubberServer.domain.admin.implement;
 
 import com.clubber.ClubberServer.domain.admin.domain.Admin;
 import com.clubber.ClubberServer.domain.admin.domain.Contact;
-import com.clubber.ClubberServer.domain.admin.domain.PendingAdminInfo;
-import com.clubber.ClubberServer.domain.admin.dto.CreateAdminSignUpRequest;
-import com.clubber.ClubberServer.domain.admin.repository.PendingAdminInfoRepository;
 import com.clubber.ClubberServer.domain.club.domain.Club;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -15,8 +12,6 @@ import org.springframework.stereotype.Component;
 public class AdminAppender {
 
     private final PasswordEncoder encoder;
-
-    private final PendingAdminInfoRepository pendingAdminInfoRepository;
 
     public void updatePassword(Admin admin, String password) {
         String encodedPassword = encoder.encode(password);
@@ -29,12 +24,6 @@ public class AdminAppender {
 
     public void updateContact(Admin admin, Contact contact) {
         admin.updateContact(contact);
-    }
-
-    public PendingAdminInfo appendPendingAdminInfo(CreateAdminSignUpRequest request) {
-        String encodedPassword = encoder.encode(request.getPassword());
-        PendingAdminInfo pendingAdminInfo = request.toEntity(encodedPassword);
-        return pendingAdminInfoRepository.save(pendingAdminInfo);
     }
 
     public void updateBySignUpApproved(Admin admin, String username, String password, Contact contact, String email) {
